+++ R/fetchNASIS_components.R
@@ -18,7 +18,7 @@
   f.comp       <- get_component_data_from_NASIS_db(SS = SS, stringsAsFactors = stringsAsFactors, dsn = dsn, nullFragsAreZero = nullFragsAreZero)
   f.chorizon   <- get_component_horizon_data_from_NASIS_db(SS = SS, fill = fill, dsn = dsn, nullFragsAreZero = nullFragsAreZero)
   f.copm       <- get_component_copm_data_from_NASIS_db(SS = SS, stringsAsFactors = stringsAsFactors, dsn = dsn)
-  f.cogeomorph <- get_component_cogeomorph_data_from_NASIS_db(SS = SS, dsn = dsn)
+  f.cogeomorph <- get_component_cogeomorph_data_from_NASIS_db2(SS = SS, dsn = dsn)
   f.otherveg   <- get_component_otherveg_data_from_NASIS_db(SS = SS, dsn = dsn)
   f.ecosite    <- get_component_esd_data_from_NASIS_db(SS = SS, stringsAsFactors = stringsAsFactors, dsn = dsn)
   f.diaghz     <- get_component_diaghz_from_NASIS_db(SS = SS, dsn = dsn)
@@ -95,12 +95,14 @@
   if (nrow(ov) > 0)
     site(f.chorizon) <- ov
 
+  # 2021-11-30: subset to hide aqp warnings for <- methods
+  
   # add diagnostic features to SPC
-  diagnostic_hz(f.chorizon) <- f.diaghz
+  diagnostic_hz(f.chorizon) <- f.diaghz[which(f.diaghz$coiid %in% f.chorizon$coiid),]
 
   # add restrictions to SPC
   # required new setter in aqp SPC object (AGB added 2019/12/23)
-  restrictions(f.chorizon) <- f.restrict
+  restrictions(f.chorizon) <- f.restrict[which(f.restrict$coiid %in% f.chorizon$coiid),]
 
   # print any messages on possible data quality problems:
   if(exists('component.hz.problems', envir=soilDB.env))