+++ tests/testthat/test-fetchSDA_component.R
@@ -47,3 +47,12 @@ test_that("fetchSDA() returns expected results", {
   # test that both components have no NRCS forest/range site assigned
   expect_equal(site(x)$ecoclassid, c(NA_character_, NA_character_))
 })
+
+test_that("fetchSDA(duplicates=TRUE) works as expected", {
+  
+  x2 <<- suppressWarnings(suppressMessages(fetchSDA(WHERE="nationalmusym = '22787'")))
+  x3 <<- suppressWarnings(suppressMessages(fetchSDA(WHERE="nationalmusym = '22787'", duplicates = TRUE)))
+ 
+  # we get a "duplicate" for each unique mukey within a nationalmusym -- 1 for each legend
+  expect_equal(length(x2) * length(unique(x3$mukey)), length(x3))
+})