+++ tests/testthat/test-waterDayYear.R
@@ -42,14 +42,12 @@ test_that("works as expected", {
   expect_true(length(x) == nrow(res))
   
   # text interface, with YYYY-MM-DD hh:mm:ss
-  res <- waterDayYear("2000-12-05 00:00:00")
+  # NOTE: waterDayYear and .formatDates involving hh:mm:ss in tests 
+  #       require `format` and `tz` argument for consistent conversion across locales
+  res <- waterDayYear("2000-12-05 12:00:00", format = "%Y-%m-%d %H:%M:%S", tz = "UTC")
   expect_equal(res$wy, 2001L)
+  
   expect_equal(res$wd, 66L)
-  
-  
-  ## what about time zones... shouldn't matter?
-  # waterDayYear(as.POSIXct("2000-12-05 00:00:00", tz = 'GMT'))
-  
   
 })